apps/api/pyproject.toml
[project]
name = "sentientops-api"
version = "0.1.0"
description = "SentientOps V1 FastAPI backend scaffold"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.16.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "psycopg[binary]>=3.2.0",
  "qdrant-client>=1.15.0",
  "mcp>=1.17.0",
  "python-dotenv>=1.1.0",
  "pyjwt[crypto]>=2.10.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0",
  "ruff>=0.12.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
